+++ pyasn1/type/univ.py
@@ -2189,15 +2189,18 @@ class SequenceAndSetBase(base.AbstractConstructedAsn1Item):
                 else:
                     myClone.setComponentByPosition(idx, componentValue.clone())
 
-    def getComponentByName(self, name):
+    def getComponentByName(self, name, instantiate=True):
         """Returns |ASN.1| type component by name.
 
         Equivalent to Python :class:`dict` subscription operation (e.g. `[]`).
 
         Parameters
         ----------
-        name : :class:`str`
+        name: :class:`str`
             |ASN.1| type component name
+
+        instantiate: :class:`bool`
+            If `True` (default), inner component will be automatically instantiated.
 
         Returns
         -------
@@ -2213,7 +2216,7 @@ class SequenceAndSetBase(base.AbstractConstructedAsn1Item):
             except KeyError:
                 raise error.PyAsn1Error('Name %s not found' % (name,))
 
-        return self.getComponentByPosition(idx)
+        return self.getComponentByPosition(idx, instantiate)
 
     def setComponentByName(self, name, value=noValue,
                            verifyConstraints=True,
@@ -2258,7 +2261,7 @@ class SequenceAndSetBase(base.AbstractConstructedAsn1Item):
             idx, value, verifyConstraints, matchTags, matchConstraints
         )
 
-    def getComponentByPosition(self, idx):
+    def getComponentByPosition(self, idx, instantiate=True):
         """Returns |ASN.1| type component by index.
 
         Equivalent to Python sequence subscription operation (e.g. `[]`).
@@ -2270,6 +2273,9 @@ class SequenceAndSetBase(base.AbstractConstructedAsn1Item):
             component or (if *componentType* is set) new ASN.1 type object gets
             instantiated.
 
+        instantiate: :class:`bool`
+            If `True` (default), inner component will be automatically instantiated.
+
         Returns
         -------
         : :py:class:`~pyasn1.type.base.PyAsn1Item`
@@ -2279,6 +2285,9 @@ class SequenceAndSetBase(base.AbstractConstructedAsn1Item):
             componentValue = self._componentValues[idx]
         except IndexError:
             componentValue = noValue
+
+        if not instantiate:
+            return componentValue
 
         if componentValue is noValue:
             self.setComponentByPosition(idx)
@@ -2538,7 +2547,7 @@ class Set(SequenceAndSetBase):
     def getComponent(self, innerFlag=False):
         return self
 
-    def getComponentByType(self, tagSet, innerFlag=False):
+    def getComponentByType(self, tagSet, instantiate=True, innerFlag=False):
         """Returns |ASN.1| type component by ASN.1 tag.
 
         Parameters
@@ -2547,13 +2556,16 @@ class Set(SequenceAndSetBase):
             Object representing ASN.1 tags to identify one of
             |ASN.1| object component
 
+        instantiate: :class:`bool`
+            If `True` (default), inner component will be automatically instantiated.
+
         Returns
         -------
         : :py:class:`~pyasn1.type.base.PyAsn1Item`
             a pyasn1 object
         """
         component = self.getComponentByPosition(
-            self.componentType.getPositionByType(tagSet)
+            self.componentType.getPositionByType(tagSet), instantiate
         )
         if innerFlag and isinstance(component, Set):
             # get inner component by inner tagSet
@@ -2733,11 +2745,11 @@ class Choice(Set):
             else:
                 myClone.setComponentByType(tagSet, component.clone())
 
-    def getComponentByPosition(self, idx):
+    def getComponentByPosition(self, idx, instantiate=True):
         __doc__ = Set.__doc__
 
         if self._currentIdx is None or self._currentIdx != idx:
-            return Set.getComponentByPosition(self, idx)
+            return Set.getComponentByPosition(self, idx, instantiate)
 
         return self._componentValues[idx]
 
@@ -2801,7 +2813,7 @@ class Choice(Set):
         else:
             return self.componentType.tagMapUnique
 
-    def getComponent(self, innerFlag=0):
+    def getComponent(self, innerFlag=False):
         """Return currently assigned component of the |ASN.1| object.
 
         Returns
@@ -2856,6 +2868,10 @@ class Choice(Set):
             return False
 
         return self._componentValues[self._currentIdx].isValue
+
+    def clear(self):
+        self._currentIdx = None
+        Set.clear(self)
 
     # compatibility stubs